request split
